[Refactoring] Try to get rid of heap allocations. Perf.
