+++ include/cocaine/trace/trace.hpp
@@ -1,5 +1,5 @@
 /*
-    Copyright (c) 2011-2015 Anton Matveenko <antmat@me.com>
+    Copyright (c) 2011-2015 Anton Matveenko <antmat@yandex-team.ru>
     Copyright (c) 2011-2015 Other contributors as noted in the AUTHORS file.
 
     This file is part of Cocaine.
@@ -14,277 +14,267 @@
     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
     GNU Lesser General Public License for more details.
 
-    You should have received a copy of the GNU Lesser Gene ral Public License
+    You should have received a copy of the GNU Lesser General Public License
     along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-#ifndef COCAINE_TRACER_TRACER_HPP
-#define COCAINE_TRACER_TRACER_HPP
-#include <blackhole/scoped_attributes.hpp>
-#include "cocaine/forwards.hpp"
+#ifndef TRACE_HPP
+#define TRACE_HPP
+
+#include "cocaine/traits/literal.hpp"
+
+#include <boost/thread/tss.hpp>
+#include <boost/optional.hpp>
 
 #include <random>
 
-namespace cocaine { namespace tracer {
+namespace cocaine {
 
-class attribute_scope_t;
-class span_t;
-class logger_t;
-
-class new_trace_scope_t;
-class trace_reset_scope_t;
-class trace_push_scope_t;
-class trace_restore_scope_t;
-
-template <class F>
-class callable_wrapper_t;
-
-typedef std::shared_ptr<span_t> span_ptr_t;
-
-/*
- ********************************************
- ************** Public API ******************
- ********************************************
-*/
-
-inline
-span_ptr_t
-current_span();
-
-inline
-void
-set_service_name(std::string name);
-
-inline void
-set_logger(std::unique_ptr<logger_t> logger);
-
-inline void
-log(std::string message);
-
-//Used to create a new trace. For example in client application.
-class new_trace_scope_t {
+// No scope here
+class empty_atttribute_scope_t {
 public:
-    inline
-    new_trace_scope_t(std::string rpc_name);
-
-    inline
-    ~new_trace_scope_t();
-private:
-    std::unique_ptr<attribute_scope_t> attr_scope;
+    //Don't do anything here by default
+    template<class ...Args>
+    empty_atttribute_scope_t(const Args& ...) {}
 };
 
-//Used to reset current trace if further tracing is undesirable - for example in logging in CNF
-class trace_reset_scope_t {
-public:
-    inline
-    trace_reset_scope_t();
+template<class ScopedAttribute>
+struct scoped_attribute_wrapper {
+    template<class Logger, class Trace>
+    scoped_attribute_wrapper(Logger& log, const Trace& t) :
+        attr(log, t.attributes())
+    {}
 
-    inline
-    ~trace_reset_scope_t();
-
-private:
-    span_ptr_t old_span;
+    template<class Logger, class Trace>
+    scoped_attribute_wrapper(Logger& log, const boost::optional<Trace>& t) :
+        attr(log, t ? t->attributes() : typename Trace::attribute_set_t())
+    {}
+    ScopedAttribute attr;
 };
 
-//Used to push a span to trace tree. Used just before making sub-RPC
-//Also holds an attribute_scope which can be used to set up some scope attributes - for example blackhole scoped attributes.
-class trace_push_scope_t {
-public:
-    // Push span
-    inline
-    trace_push_scope_t(std::string rpc_name);
-
-    // Push span and log an anotation via tracer::log
-    inline
-    trace_push_scope_t(std::string annotation, std::string rpc_name);
-
-    inline
-    ~trace_push_scope_t();
-private:
-    std::unique_ptr<attribute_scope_t> attr_scope;
+template<>
+struct scoped_attribute_wrapper<empty_atttribute_scope_t> {
+    template<class Logger, class Trace>
+    scoped_attribute_wrapper(const Logger&, const Trace&) {}
 };
 
-// Restores trace from some external source.
-// For example when message was read from network or
-// we executing some async operation with previously stored span
-// Can be default constructed and restored later only once OR non default constructed.
-// calling restore after non default ctor or previous restore is misusage.
-// When object of this type goes out of scope it restores original trace
-class trace_restore_scope_t {
+template<class Logger, class ScopedAttribute, class AttributeSet>
+class trace {
 public:
-    inline
-    trace_restore_scope_t() = default;
+    typedef AttributeSet attribute_set_t;
+    template<size_t N>
+    struct stack_str_t {
+        static constexpr size_t max_size = N;
+        char blob[N+1];
+        size_t size;
 
-    // Restores trace from trace-span-parent param. Logs event via logger.
-    inline
-    trace_restore_scope_t(std::string annotation, std::string rpc_name, uint64_t trace_id, uint64_t span_id, uint64_t parent_id);
+        stack_str_t(const char* lit) {
+            memcpy(blob, lit, std::min(N, strlen(lit)));
+            blob[N] = '\0';
+        }
 
-    // Restores trace from saved span. Logs annotation via logger.
-    inline
-    trace_restore_scope_t(std::string annotation, span_ptr_t span);
+        stack_str_t(const char* lit, size_t sz) {
+            memcpy(blob, lit, std::min(N, sz));
+            blob[N] = '\0';
+        }
 
-    // Restores trace from saved span. No-log.
-    inline
-    trace_restore_scope_t(span_ptr_t span);
+        stack_str_t(const std::string& source) {
+            memcpy(blob, source.c_str(), std::min(N, source.size()));
+            blob[N] = '\0';
+        }
 
-    // Restores trace from saved span. Logs annotation via logger.
-    inline
-    void
-    restore(std::string annotation, span_ptr_t span);
+        stack_str_t() :
+            blob(),
+            size(0)
+        {}
+    };
 
-    // Restores trace from saved span. No-log.
-    inline
-    void
-    restore(span_ptr_t span);
+    template <class F>
+    class callable_wrapper_t;
 
-    // Restores trace from trace-span-parent param. Logs event via logger.
-    inline
-    void
-    restore(std::string annotation, std::string rpc_name, uint64_t trace_id, uint64_t span_id, uint64_t parent_id);
+    class restore_scope_t {
+    public:
+        restore_scope_t(boost::optional<trace> new_trace) :
+            old_span(),
+            attribute(*trace::get_logger(), new_trace)
+        {
+            if(new_trace) {
+                old_span = trace::current();
+                trace::current() = new_trace.get();
+            }
+        }
+        ~restore_scope_t() {
+            if(old_span) {
+                trace::current() = old_span.get();
+            }
+        }
+    private:
+        boost::optional<trace> old_span;
+        scoped_attribute_wrapper<ScopedAttribute> attribute;
+    };
 
-    inline
-    ~trace_restore_scope_t();
+    class push_scope_t {
+    public:
+        template<class ServiceStr, class RpcStr>
+        push_scope_t(const RpcStr& _rpc_name, const ServiceStr& _service_name):
+            attribute(*trace::get_logger(), push())
+        {
+        }
 
-private:
-    span_ptr_t old_span;
-    std::unique_ptr<attribute_scope_t> attr_scope;
-};
+        ~push_scope_t() {
+            trace::current().pop();
+        }
 
-//Bind callback to store trace and log now and on callback call.
-template<class... Args>
-auto
-bind(std::string message, Args&& ...args) -> callable_wrapper_t<decltype(std::bind(std::forward<Args>(args)...))>;
+    private:
+        trace&
+        push() {
+            return trace::current().push();
+        }
 
-//Bind callback to store trace. Behaves exactly as std::bind
-template<class... Args>
-auto
-bind(Args&& ...args) -> callable_wrapper_t<decltype(std::bind(std::forward<Args>(args)...))>;
+        scoped_attribute_wrapper<ScopedAttribute> attribute;
+    };
 
-inline uint64_t generate_id();
+    trace() :
+        trace_id(),
+        span_id(),
+        parent_id(),
+        parent_parent_id(),
+        start_time_us(),
+        last_time_us(),
+        rpc_name(),
+        service_name()
+    {}
 
-class span_t :
-    public std::enable_shared_from_this<span_t>
-{
-public:
-    inline
-    uint64_t
-    get_parent_id() const {
-        return parent && !parent->empty() ? parent->span_id : 0;
+    /*
+    trace(const trace& other) :
+        trace_id(other.trace_id),
+        span_id(other.span_id),
+        parent_id(other.parent_id),
+        parent_parent_id(other.parent_id),
+        start_time_us(other.start_time_us),
+        last_time_us(other.last_time_us),
+        name(other.name)
+    {}
+    */
+
+    template<class ServiceStr, class RpcStr>
+    trace(uint64_t _trace_id,
+          uint64_t _span_id,
+          uint64_t _parent_id,
+          const RpcStr& _rpc_name,
+          const ServiceStr& _service_name) :
+        trace_id(_trace_id),
+        span_id(_span_id),
+        parent_id(_parent_id),
+        parent_parent_id(),
+        start_time_us(),
+        last_time_us(),
+        rpc_name(_rpc_name),
+        service_name(_service_name)
+    {
     }
 
-    inline
+    static
+    trace&
+    current() {
+        static boost::thread_specific_ptr<trace> t;
+        if(t.get() == nullptr) {
+            t.reset(new trace());
+        }
+        return *t.get();
+    }
+
+    static
+    Logger*
+    get_logger() {
+        return logger;
+    }
+
+    static
+    void
+    set_logger(Logger* _logger) {
+        logger = _logger;
+    }
+
+//    static
+//    void
+//    set_default_service_name(const literal_t& new_name) {
+//        default_service_name = new_name;
+//    }
+
+    template<class... Args>
+    static
+    auto
+    bind(Args&& ...args) -> callable_wrapper_t<decltype(std::bind(std::forward<Args>(args)...))>;
+
+    template<class Method>
+    static
+    auto
+    mem_fn(Method m) -> callable_wrapper_t<decltype(std::mem_fn(std::forward<Method>(m)))>;
+
+    uint64_t
+    get_parent_id() const {
+        return parent_id;
+    }
+
     uint64_t
     get_trace_id() const {
         return trace_id;
     }
 
-    inline
     uint64_t
     get_id() const {
         return span_id;
     }
 
-    inline
-    blackhole::attribute::set_t
-    attributes() const;
-
-    static
-    inline
-    uint64_t
-    cur_time(){
-        return std::chrono::duration_cast<std::chrono::microseconds>(
-            std::chrono::system_clock::now().time_since_epoch()
-        ).count();
+    trace& pop() {
+        assert(parent_parent_id != 0);
+        span_id = parent_id;
+        parent_id = parent_parent_id;
+        parent_parent_id = 0;
     }
 
-    inline
-    bool
-    empty() const {
-        return span_id == 0;
+    trace& push() {
+        parent_parent_id = parent_id;
+        parent_id = span_id;
+        span_id = generate_id();
     }
 
-    inline
-    const std::string&
-    get_name() const {
-        return name;
+    uint64_t generate_id() {
+        static std::random_device rd;
+        static std::mt19937 gen(rd());
+        static std::uniform_int_distribution<uint64_t> dis(1, std::numeric_limits<uint64_t>::max()/2-1);
+        return dis(gen);
     }
 
+    AttributeSet
+    attributes() const {
+        return {
+            {"trace_id", {trace_id}},
+            {"span_id", {trace_id}},
+            {"parent_id", {trace_id}},
+            {"rpc_name", {rpc_name.blob}},
+            {"service_name", {service_name.blob}}
+        };
+    }
 private:
-    span_t() :
-        trace_id(),
-        span_id(),
-        start_time_us(),
-        last_time_us(),
-        parent(nullptr),
-        name()
-    {}
-
-    span_t(std::string _name, uint64_t _trace_id, uint64_t _span_id, uint64_t _parent_id) :
-        trace_id(_trace_id ? _trace_id : generate_id()),
-        span_id(_span_id ? _span_id : trace_id),
-        start_time_us(cur_time()),
-        last_time_us(start_time_us),
-        parent(new span_t("", trace_id, _parent_id)),
-        name(_name)
-    {}
-
-
-    span_t(std::string _name, std::shared_ptr<span_t> _parent) :
-        trace_id(_parent->empty() ? generate_id() : _parent->trace_id),
-        span_id(_parent->empty() ? trace_id : generate_id()),
-        start_time_us(cur_time()),
-        last_time_us(start_time_us),
-        parent(std::move(_parent)),
-        name(_name)
-    {}
-
-    inline
-    void
-    set_name(std::string new_name) {
-        last_time_us = cur_time();
-        name = std::move(new_name);
-    }
-
-    inline
-    uint64_t
-    get_last_time() const {
-        return last_time_us;
-    }
-
-    inline
-    uint64_t
-    duration() const {
-        return cur_time() - last_time_us;
-    }
-
-    inline
-    uint64_t
-    full_duration() const {
-        return cur_time() - start_time_us;
-    }
-
-    span_t(std::string _name, uint64_t _trace_id, uint64_t _span_id) :
-        trace_id(_trace_id),
-        span_id(_span_id),
-        start_time_us(cur_time()),
-        last_time_us(start_time_us),
-        parent(nullptr),
-        name(_name)
-    {}
-
     uint64_t trace_id;
     uint64_t span_id;
+    uint64_t parent_id;
+    uint64_t parent_parent_id;
     uint64_t start_time_us;
     uint64_t last_time_us;
-    std::shared_ptr<span_t> parent;
-    std::string name;
+    stack_str_t<16> rpc_name;
+    stack_str_t<16> service_name;
 
-    friend class trace_context_t;
-    friend class trace_reset_scope_t;
+    static Logger* logger;
 };
 
-}}
+template<class Logger, class ScopedAttribute, class AttributeSet>
+Logger* trace<Logger, ScopedAttribute, AttributeSet>::logger = nullptr;
 
+}
 #include "cocaine/trace/trace_impl.hpp"
 
-#endif
+#endif // TRACE_HPP
+
